# apps/daemon/openapi.json
{
  "components": {
    "schemas": {
...
      "AcceptBody": {
        "properties": {
          "new_text": {
            "anyOf": [
              {
                "maxLength": 200000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Text"
          }
        },
        "title": "AcceptBody",
        "type": "object"
      },
...
      "BatchBody": {
        "properties": {
          "ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "minItems": 1,
            "title": "Ids",
            "type": "array"
          }
        },
        "required": [
          "ids"
        ],
        "title": "BatchBody",
        "type": "object"
      },
      "BatchResult": {
        "properties": {
          "applied": {
            "items": {
              "type": "string"
            },
            "title": "Applied",
            "type": "array"
          },
          "stopped_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stopped At"
          }
        },
        "required": [
          "applied"
        ],
        "title": "BatchResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OptInBody": {
        "properties": {
          "source": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "source"
        ],
        "title": "OptInBody",
        "type": "object"
      },
...
      "PageProperty": {
        "properties": {
          "colors": {
            "additionalProperties": {
              "enum": [
                "default",
                "gray",
                "brown",
                "orange",
                "yellow",
                "green",
                "blue",
                "purple",
                "pink",
                "red"
              ],
              "type": "string"
            },
            "maxProperties": 100,
            "title": "Colors",
            "type": "object"
          },
          "id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Options",
            "type": "array"
          },
          "type": {
            "enum": [
              "text",
              "number",
              "single_select",
              "multi_select",
              "date",
              "checkbox",
              "email",
              "url",
              "media",
              "status",
              "created",
              "updated"
            ],
            "title": "Type",
            "type": "string"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "id",
          "name",
          "type"
        ],
        "title": "PageProperty",
        "type": "object"
      },
...
      "Proposal": {
        "properties": {
          "applied_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied Hash"
          },
          "base_body_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Body Hash"
          },
          "base_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Hash"
          },
          "base_properties": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PageProperty"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Properties"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By"
          },
          "edited": {
            "default": false,
            "title": "Edited",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "new_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Path"
          },
          "new_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Text"
          },
          "old_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Old Text"
          },
          "opt_in_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opt In Source"
          },
          "page_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Id"
          },
          "page_path": {
            "title": "Page Path",
            "type": "string"
          },
          "page_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Title"
          },
          "parent_proposal_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Proposal Id"
          },
          "patch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patch"
          },
          "policy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy"
          },
          "properties": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PageProperty"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Properties"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "snapshot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "trash_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trash Id"
          }
        },
        "required": [
          "id",
          "page_path",
          "kind",
          "status",
          "created_at"
        ],
        "title": "Proposal",
        "type": "object"
      },
...
      "RejectBatchBody": {
        "properties": {
          "ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "minItems": 1,
            "title": "Ids",
            "type": "array"
          },
          "reason": {
            "default": "",
            "maxLength": 1000,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "ids"
        ],
        "title": "RejectBatchBody",
        "type": "object"
      },
      "RejectBatchResult": {
        "properties": {
          "rejected": {
            "items": {
              "type": "string"
            },
            "title": "Rejected",
            "type": "array"
          }
        },
        "required": [
          "rejected"
        ],
        "title": "RejectBatchResult",
        "type": "object"
      },
      "RejectBody": {
        "properties": {
          "reason": {
            "default": "",
            "maxLength": 1000,
            "title": "Reason",
            "type": "string"
          }
        },
        "title": "RejectBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "graite__api__review__OptInResult": {
        "properties": {
          "opted_in": {
            "items": {
              "type": "string"
            },
            "title": "Opted In",
            "type": "array"
          }
        },
        "required": [
          "opted_in"
        ],
        "title": "OptInResult",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "graite-daemon",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ai/proposals": {
      "get": {
        "operationId": "list_proposals_api_v1_ai_proposals_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "conversation_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Conversation Id"
            }
          },
          {
            "in": "query",
            "name": "page_path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page Path"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Proposal"
                  },
                  "title": "Response List Proposals Api V1 Ai Proposals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Proposals",
        "tags": [
          "AI"
        ]
      }
    },
    "/api/v1/ai/proposals/accept-batch": {
      "post": {
        "operationId": "accept_batch_api_v1_ai_proposals_accept_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Batch",
        "tags": [
          "AI"
        ]
      }
    },
    "/api/v1/ai/proposals/opt-in": {
      "post": {
        "description": "Confirm once that a folder's auto-apply setting may apply changes without review.",
        "operationId": "opt_in_auto_apply_api_v1_ai_proposals_opt_in_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptInBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/graite__api__review__OptInResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Opt In Auto Apply",
        "tags": [
          "AI"
        ]
      }
    },
    "/api/v1/ai/proposals/reject-batch": {
      "post": {
        "description": "Discard several open proposals at once. Without a reason nothing is fed back to the model.",
        "operationId": "reject_batch_api_v1_ai_proposals_reject_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectBatchBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectBatchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Batch",
        "tags": [
          "AI"
        ]
      }
    },
    "/api/v1/ai/proposals/{proposal_id}": {
      "get": {
        "operationId": "get_proposal_api_v1_ai_proposals__proposal_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Proposal"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Proposal",
        "tags": [
          "AI"
        ]
      }
    },
    "/api/v1/ai/proposals/{proposal_id}/accept": {
      "post": {
        "operationId": "accept_proposal_api_v1_ai_proposals__proposal_id__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Proposal"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Proposal",
        "tags": [
          "AI"
        ]
      }
    },
    "/api/v1/ai/proposals/{proposal_id}/reject": {
      "post": {
        "operationId": "reject_proposal_api_v1_ai_proposals__proposal_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Proposal"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Proposal",
        "tags": [
          "AI"
        ]
      }
    },
    "/api/v1/ai/proposals/{proposal_id}/revert": {
      "post": {
        "operationId": "revert_proposal_api_v1_ai_proposals__proposal_id__revert_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Proposal"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revert Proposal",
        "tags": [
          "AI"
        ]
      }
    },
...
  }
}